requirements.txt
Django==3.2.10
djangorestframework==3.12.4
djongo==1.3.6
pymongo==3.12.3
neo4j==4.4.0
python-dotenv==0.19.2
django-cors-headers==3.10.1
